Copying default configs to new cluster deployment. Setting node_type in script configs

+++ dev-ops/start-es-cluster-v6-local_drive.sh
@@ -2,12 +2,15 @@ dir=`pwd`
 cd
 
 STACK_DIR=~/workspace/elastic-stack/current
+PROJECT_STACK=~/workspace/tesla-stack
 ES_HOME=$STACK_DIR/elasticsearch
 CLUSTER_NAME=local_data.2019-01-07
+
 if [ -z $ENV_CLUSTER_NAME ] ; then
   echo "Setting cluster.name with env vare: $ENV_CLUSTER_NAME"
   CLUSTER_NAME=$ENV_CLUSTER_NAME
 fi
+
 DATA_DRIVE=/data/local_drive
 CLUSTER_DIR=$DATA_DRIVE/elasticsearch/$CLUSTER_NAME
 
@@ -15,12 +18,16 @@ DATA_DIR=$CLUSTER_DIR/data
 PATH_CONFIGS=$CLUSTER_DIR/configs
 PIDS_DIR=$CLUSTER_DIR/pids
 LOGS_DIR=$CLUSTER_DIR/logs
-PROJECT_STACK=~/workspace/tesla-stack
+
 DEFAULT_CONFIGS_PATH_ES=${PROJECT_STACK}/default-configs/elasticsearch
 DEFAULT_NODE_TYPE=default.config
+HOT_NODE_TYPE=hot.config
+WARM_NODE_TYPE=warm.config
+COOL_NODE_TYPE=cool.config
 NUM_HOT_NODES=1
 NUM_WARM_NODES=1
 NUM_COOL_NODES=1
+NODE_HEAP_GB=4
 
 echo "Checking that directory path exists: $DATA_DIR"
 # Set up $DATA_DIR paths
@@ -54,8 +61,8 @@ if [ ! -z $PATH_CONFIGS ]; then
 # Need root privs to create data, config, log paths
   sudo mkdir -p $PATH_CONFIGS
   echo "Created pids directory: $PATH_CONFIGS"
-  echo "TODO: Copy defaults"
 fi
+
 
 # --- Start Elasticsearch Processes ---
 if [ -z ES_HOME ]; then 
@@ -64,7 +71,9 @@ else cd ~/workspace/elastic-stack/current/elasticsearch
 fi
 
 # NODE 1
-NODE_NAME=node-d0
+NODE_NAME=node-0
+NODE_HEAP_GB=4
+NODE_TYPE=$HOT_NODE_TYPE
 if [ ! -z $DATA_DIR/$NODE_NAME ]; then
   sudo mkdir -p $DATA_DIR/$NODE_NAME
   echo "Created node data directory: $DATA_DIR/$NODE_NAME"
@@ -73,11 +82,22 @@ fi
 if [ ! -z $PATH_CONFIGS/$NODE_NAME  ]; then
   sudo mkdir -p $PATH_CONFIGS/$NODE_NAME
   echo "Created node config directory: $PATH_CONFIGS/$NODE_NAME"
-  echo "Copying default configs from: $EFAULT_CONFIGS_PATH_ES/$DEFAULT_NODE_TYPE"
-  sudo cp -rf $DEFAULT_CONFIGS_PATH_ES/$DEFAULT_NODE_TYPE/* $PATH_CONFIGS/$NODE_NAME
+  echo "Copying default configs from: $DEFAULT_CONFIGS_PATH_ES/$NODE_TYPE"
+  if [ ! -f $PATH_CONFIGS/$NODE_NAME/elasticsearch.yml ]; then 
+    echo "HERE $PATH_CONFIGS/$NODE_NAME/elasticsearch.yml better not exist!"
+    sudo cp -rf $DEFAULT_CONFIGS_PATH_ES/$DEFAULT_NODE_TYPE/* $PATH_CONFIGS/$NODE_NAME
+  fi
 fi
 
-ES_JAVA_OPTS="-Xms4g -Xmx4g" ES_PATH_CONF="${PATH_CONFIGS}/${NODE_NAME}" ./bin/elasticsearch -p ${PIDS_DIR}/${NODE_NAME}.pid -E path.data=${DATA_DIR}/${NODE_NAME}  -E path.logs=${LOGS_DIR}  -E "cluster.name=${CLUSTER_NAME}" -E "node.name"=${NODE_NAME}  -d
+
+# IMPORTANT - chown node directories to non-root user
+echo "IMPORTANT: chowning recursively all cluster config, data directories created before starting up cluster.."
+sudo chown -R andrew:wheel ${PATH_CONFIGS}/${NODE_NAME}
+sudo chown -R andrew:wheel ${DATA_DIR}/${NODE_NAME}
+sudo chown -R andrew:wheel ${LOGS_DIR}/$NODE_NAME
+
+
+ES_JAVA_OPTS="-Xms4g -Xmx4g" ES_PATH_CONF="${PATH_CONFIGS}/${NODE_NAME}" ./bin/elasticsearch -p ${PIDS_DIR}/${NODE_NAME}.pid -E path.data=${DATA_DIR}/${NODE_NAME}  -E path.logs=${LOGS_DIR}/$NODE_NAME  -E "cluster.name=${CLUSTER_NAME}" -E node.name=${NODE_NAME}  -d
 echo "Started Elasticsearch $NODE_NAME"
 echo "Logs are at $LOGS_DIR"
 echo "Data is at $DATA_DIR"
@@ -85,7 +105,9 @@ echo "Configs created are ate $PATHS_CONFIG/$NODE_NAME"
 
 # NODE 2
 
-NODE_NAME=node-d1
+NODE_NAME=node-1
+NODE_HEAP_GB=2
+NODE_TYPE=$COOL_NODE_TYPE
 if [ ! -z $DATA_DIR/$NODE_NAME ]; then
   mkdir -p $DATA_DIR/$NODE_NAME
   echo "Created node data directory: $DATA_DIR/$NODE_NAME"
@@ -94,15 +116,22 @@ fi
 if [ ! -z $PATH_CONFIGS/$NODE_NAME  ]; then
   sudo mkdir -p $PATH_CONFIGS/$NODE_NAME
   echo "Created node config directory: $PATH_CONFIGS/$NODE_NAME"
-  echo "Copying default configs from: $EFAULT_CONFIGS_PATH_ES/$DEFAULT_NODE_TYPE"
-  sudo cp -rf $DEFAULT_CONFIGS_PATH_ES/$DEFAULT_NODE_TYPE/* $PATH_CONFIGS/$NODE_NAME
+  echo "Copying default configs from: $DEFAULT_CONFIGS_PATH_ES/$NODE_TYPE"
+  if [ ! -f $PATH_CONFIGS/$NODE_NAME/elasticsearch.yml ]; then 
+    sudo cp -rf $DEFAULT_CONFIGS_PATH_ES/$DEFAULT_NODE_TYPE/* $PATH_CONFIGS/$NODE_NAME
+  fi
 fi
 if [ -z ES_HOME ]; then 
   cd $ES_HOME
 else cd ~/workspace/elastic-stack/current/elasticsearch
 fi
 
-ES_JAVA_OPTS="-Xms2g -Xmx2g" ES_PATH_CONF="${PATH_CONFIGS}/${NODE_NAME}" ./bin/elasticsearch -E path.data=${DATA_DIR}/${NODE_NAME}  -p ${PIDS_DIR}/${NODE_NAME}.pid -E path.logs=${LOGS_DIR} -E cluster.name=${CLUSTER_NAME}  -E "node.name"=${NODE_NAME} -d
+# IMPORTANT - chown node directories to non-root user
+echo "IMPORTANT: chowning recursively all cluster config, data directories created before starting up cluster.."
+sudo chown -R andrew:wheel ${PATH_CONFIGS}/${NODE_NAME}
+sudo chown -R andrew:wheel ${DATA_DIR}/${NODE_NAME}
+sudo chown -R andrew:wheel ${LOGS_DIR}/$NODE_NAME
+ES_JAVA_OPTS="-Xms2g -Xmx2g" ES_PATH_CONF="${PATH_CONFIGS}/${NODE_NAME}" ./bin/elasticsearch -E path.data=${DATA_DIR}/${NODE_NAME}  -p ${PIDS_DIR}/${NODE_NAME}.pid -E path.logs=${LOGS_DIR}/$NODE_NAME -E cluster.name=${CLUSTER_NAME}  -E node.name=${NODE_NAME} -d
 echo "Started Elasticsearch $NODE_NAME"
 echo "Logs are at $LOGS_DIR"
 echo "Data is at $DATA_DIR"
@@ -110,24 +139,32 @@ echo "Configs created are ate $PATHS_CONFIG/$NODE_NAME"
 
 # NODE 3
 
-NODE_NAME=node-d2
+NODE_NAME=node-2
+NODE_HEAP_GB=2
+NODE_TYPE=$COOL_NODE_TYPE
 if [ ! -z $DATA_DIR/$NODE_NAME ]; then
   mkdir -p $DATA_DIR/$NODE_NAME
   echo "Created node data directory: $DATA_DIR/$NODE_NAME"
-  echo "Copying default configs from: $EFAULT_CONFIGS_PATH_ES/$DEFAULT_NODE_TYPE"
-  sudo cp -rf $DEFAULT_CONFIGS_PATH_ES/$DEFAULT_NODE_TYPE/* $PATH_CONFIGS/$NODE_NAME
 fi
 # Create node config directories, if they don't exist
 if [ ! -z $PATH_CONFIGS/$NODE_NAME  ]; then
   sudo mkdir -p $PATH_CONFIGS/$NODE_NAME
   echo "Created node config directory: $PATH_CONFIGS/$NODE_NAME"
+  echo "Copying default configs from: $DEFAULT_CONFIGS_PATH_ES/$NODE_TYPE"
+  if [ ! -f $PATH_CONFIGS/$NODE_NAME/elasticsearch.yml ]; then 
+    sudo cp -rf $DEFAULT_CONFIGS_PATH_ES/$DEFAULT_NODE_TYPE/$NODE_TYPE $PATH_CONFIGS/$NODE_NAME
+  fi 
 fi
 if [ -z ES_HOME ]; then 
   cd $ES_HOME
 else cd ~/workspace/elastic-stack/current/elasticsearch
 fi
-echo "HERE `pwd`"
-ES_JAVA_OPTS="-Xms2g -Xmx2g" ES_PATH_CONF="${PATH_CONFIGS}/${NODE_NAME}" ./bin/elasticsearch -p ${PIDS_DIR}/${NODE_NAME}.pid -E path.data=${DATA_DIR}/${NODE_NAME}  -E path.logs=${LOGS_DIR} -E node.name=${NODE_NAME} -E "cluster.name=${CLUSTER_NAME}" -d
+# IMPORTANT - chown node directories to non-root user
+echo "IMPORTANT: chowning recursively all cluster config, data directories created before starting up cluster.."
+sudo chown -R andrew:wheel ${PATH_CONFIGS}/${NODE_NAME}
+sudo chown -R andrew:wheel ${DATA_DIR}/${NODE_NAME}
+sudo chown -R andrew:wheel ${LOGS_DIR}/$NODE_NAME
+ES_JAVA_OPTS="-Xms1g -Xmx1g" ES_PATH_CONF="${PATH_CONFIGS}/${NODE_NAME}" ./bin/elasticsearch -p ${PIDS_DIR}/${NODE_NAME}.pid -E path.data=${DATA_DIR}/${NODE_NAME}  -E path.logs=${LOGS_DIR}/$NODE_NAME -E node.name=${NODE_NAME} -E cluster.name=${CLUSTER_NAME} -d
 echo "Started Elasticsearch $NODE_NAME" 
 echo "Logs are at $LOGS_DIR"
 echo "Data is at $DATA_DIR"
